requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
backoff==2.2.1
beautifulsoup4==4.14.3
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
cymem==2.0.13
dataclasses-json==0.6.7
Deprecated==1.3.1
effdet==0.4.1
einops==0.8.2
emoji==2.15.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl
es_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.8.0/es_core_news_lg-3.8.0-py3-none-any.whl
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.133.0
filelock==3.24.3
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-vision==3.12.1
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.78.1
h11==0.16.0
h2==4.3.0
hf-xet==1.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.4.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
langdetect==1.0.9
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
murmurhash==1.0.15
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.3
numba==0.64.0
numpy==2.2.6
olefile==0.47
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.23.2
opencv-python==4.13.0.92
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
phonenumbers==9.0.24
pi_heif==1.2.1
pikepdf==10.3.0
pillow==12.1.1
portalocker==3.2.0
preshed==3.0.12
presidio_analyzer==2.2.361
presidio_anonymizer==2.2.361
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycocotools==2.0.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.27.1
pyparsing==3.3.2
pypdf==6.7.2
pypdfium2==5.5.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.17.0
RapidFuzz==3.14.3
regex==2026.2.19
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
schedule==1.2.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.3
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
soupsieve==2.8.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.52.1
sympy==1.14.0
thinc==8.3.10
threadpoolctl==3.6.0
timm==1.0.25
tldextract==5.3.1
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.18.32
unstructured-client==0.42.10
unstructured.pytesseract==0.3.15
unstructured_inference==1.2.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.3
webencodings==0.5.1
wrapt==2.1.1
xlsxwriter==3.2.9
